Refactor board setup
