+++ src/minesweeper/board/boardState.svelte.ts
@@ -1,44 +1,3 @@
-import { BOARD_SIZE, MINE_AMOUNT, DIRECTIONS } from "../constants";
-import type { Cell } from "../types";
-
-const setup = (): Cell[][] => {
-  const board = Array.from(Array(BOARD_SIZE.ROWS), () =>
-    Array.from(
-      Array(BOARD_SIZE.COLS),
-      (): Cell => ({
-        content: "empty",
-        status: "initial",
-        bombsAround: 0,
-      }),
-    ),
-  );
-
-  function fillBombCount(row: number, col: number) {
-    if (row < 0 || row >= board.length || col < 0 || col >= board[0].length) {
-      return;
-    }
-
-    board[row][col].bombsAround++;
-  }
-
-  for (let i = 0; i < MINE_AMOUNT; i++) {
-    const randomRow = Math.floor(Math.random() * BOARD_SIZE.ROWS);
-    const randomCol = Math.floor(Math.random() * BOARD_SIZE.COLS);
-    const cell = board[randomRow][randomCol];
-
-    if (cell.content === "mine") {
-      i--;
-      continue;
-    }
-
-    cell.content = "mine";
-
-    for (const [rowOffset, colOffset] of DIRECTIONS) {
-      fillBombCount(randomRow + rowOffset, randomCol + colOffset);
-    }
-  }
-
-  return board;
-};
+import { setup } from "./boardSetup";
 
 export const boardState = $state(setup());
